fix: remove deps, etc
