+++ .circleci/config.yml
@@ -1,7 +1,7 @@
 defaults: &defaults
   working_directory: ~/code
   docker:
-    - image: circleci/node:12
+    - image: circleci/node:16
       environment:
         NPM_CONFIG_LOGLEVEL: error # make npm commands less noisy
         JOBS: max # https://gist.github.com/ralphtheninja/f7c45bdee00784b41fed
@@ -9,11 +9,11 @@ defaults: &defaults
 restore_cache: &restore_cache
   restore_cache:
     keys:
-      - code-{{ .Branch }}-{{ checksum ".nvmrc" }}-{{ checksum "yarn.lock" }}
+      - code-{{ .Branch }}-{{ checksum ".nvmrc" }}-{{ checksum "package-lock.json" }}
 
 save_cache: &save_cache
   save_cache:
-      key: code-{{ .Branch }}-{{ checksum ".nvmrc" }}-{{ checksum "yarn.lock" }}
+      key: code-{{ .Branch }}-{{ checksum ".nvmrc" }}-{{ checksum "package-lock.json" }}
       paths:
         - node_modules
 
@@ -24,10 +24,10 @@ jobs:
     steps:
       - checkout
       - *restore_cache
-      - run: yarn
-      - run: yarn build
-      - run: yarn tslint
-      - run: yarn test
+      - run: npm install
+      - run: npm run build
+      - run: npm run lint
+      - run: npm run test
       - run: npx codecov
       - *save_cache
