+++ .github/dependabot.yml
@@ -6,17 +6,3 @@ updates:
     interval: daily
     time: "13:00"
   open-pull-requests-limit: 2
-  ignore:
-  - dependency-name: y18n
-    versions:
-    - 4.0.1
-    - 4.0.2
-  - dependency-name: husky
-    versions:
-    - 4.3.8
-    - 5.0.9
-    - 5.1.0
-    - 5.1.1
-    - 5.1.2
-    - 5.1.3
-    - 5.2.0
