+++ .gitignore
@@ -4,7 +4,7 @@ logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-package-lock.json
+yarn.lock
 
 # Runtime data
 pids
@@ -60,3 +60,5 @@ typings/
 
 # next.js build output
 .next
+
+.dccache
