--- .markdownlintrc
@@ -1,14 +1,0 @@
-{
-  "default": true,
-  "MD007": {
-    "indent": 2
-  },
-  "line-length": false,
-  "first-line-h1": false,
-  "first-header-h1": false,
-  "no-duplicate-header": false,
-  "no-inline-html": false,
-  "no-hard-tabs": false,
-  "single-h1": false,
-  "whitespace": false
-}
