--- .spelling
@@ -1,11 +1,0 @@
-# markdown-spellcheck spelling configuration file
-# Format - lines beginning # are comments
-# global dictionary is at the start, file overrides afterwards
-# one word per line, to define a file override use ' - filename'
-# where filename is relative to this configuration file
-intl
- - README.md
-regex
-IEs
-tolocalestring
-IE11
