+++ index.d.ts
@@ -1,3 +1,2 @@
 export declare const generateCleanNumber: (num: string | number) => number;
 export declare const filterNaNfromNumber: (num: string | number) => number;
-//# sourceMappingURL=index.d.ts.map
