+++ index.js
@@ -1,11 +1,13 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.generateCleanNumber = function (num) {
+exports.filterNaNfromNumber = exports.generateCleanNumber = void 0;
+var generateCleanNumber = function (num) {
     return typeof num !== "string" || isNaN(Number(num))
         ? parseFloat(num.toString().replace(/[^0-9,.]/g, ""))
         : parseFloat(num);
 };
-exports.filterNaNfromNumber = function (num) {
+exports.generateCleanNumber = generateCleanNumber;
+var filterNaNfromNumber = function (num) {
     if (typeof num === "number")
         return num;
     var numArray = num.split("");
@@ -13,4 +15,4 @@ exports.filterNaNfromNumber = function (num) {
     var filteredNumString = filteredNumArray.join("");
     return parseFloat(filteredNumString);
 };
-//# sourceMappingURL=index.js.map
+exports.filterNaNfromNumber = filterNaNfromNumber;
