+++ package.json
@@ -1,56 +1,32 @@
 {
   "name": "generate-clean-number",
-  "version": "3.0.0",
+  "version": "3.0.1",
   "description": "A JavaScript utility for removing anything that is not a number (NaN) from a number",
   "main": "index.js",
   "repository": "git@github.com:yowainwright/generate-clean-number.git",
   "author": "Jeff Wainwright <yowainwright@gmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "@types/jest": "^27.0.1",
     "@types/node": "^17.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.27.0",
+    "@typescript-eslint/parser": "^5.27.0",
     "codecov": "^3.2.0",
-    "husky": "^7.0.1",
-    "jest": "^26.0.1",
-    "markdown-spellcheck": "^1.3.1",
-    "markdownlint-cli": "^0.31.1",
-    "ts-jest": "^26.1.1",
-    "tslint": "^6.1.2",
+    "eslint": "^8.17.0",
+    "eslint-config-prettier": "^8.5.0",
     "typescript": "^4.3.5",
-    "write-good": "^1.0.1"
+    "vitest": "^0.14.1"
   },
   "scripts": {
     "build": "tsc",
-    "markdownlint": "markdownlint *.md",
     "prepush": "npm run build && npm test",
-    "pre-commit-msg": "Echo 'Running pre-commit checks...' && exit 0",
-    "postpublish": "git tag $npm_package_version && git push origin --tags",
-    "release": "npm run lint && npm run build && npm test && npm run report:coverage",
-    "release-ready": "npm run lint && npm run build && npm test",
     "report:coverage": "nyc report --reporter=lcov > coverage.lcov && codecov",
-    "spelling": "mdspell '**/*.md' '!**/node_modules/**/*.md' --ignore-numbers",
-    "spelling:ci": "mdspell '**/*.md' '!**/node_modules/**/*.md' --ignore-numbers --report",
     "ts-check": "tsc --noEmit",
-    "tslint": "tslint -c tslint.json .",
+    "lint": "eslint index.ts",
     "start": "npm install",
-    "test": "jest --coverage"
+    "test": "vitest run"
   },
   "files": [
     "index.js",
     "index.d.ts"
-  ],
-  "jest": {
-    "testURL": "https://localhost/",
-    "transform": {
-      "^.+\\.ts?$": "ts-jest"
-    },
-    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(ts?)$",
-    "moduleFileExtensions": [
-      "ts",
-      "js"
-    ]
-  },
-  "quokka": {
-    "babel": true
-  }
+  ]
 }
